新增发布新文章功能demo

+++ src/main/java/com/murasame/domain/vo/BlogBriefVO.java
@@ -8,7 +8,7 @@ import java.time.LocalDateTime;
 // 博客摘要词条 - 前端展示VO
 @Data
 public class BlogBriefVO {
-	private BigInteger id;              // blogs.id
+	private Long id;              // blogs.id
 	private String title;               // blogs.title
 	private String brief;               // top 30 char of blogs.content
 	private LocalDateTime created_at;   // same
